requirements.txt
Flask==3.0.0
Pillow==9.2.0
tensorflow==2.15.0
keras==3.1.1
